Pin dependency versions in the Dockerfile
This is the last step to give more repeatable builds. This ensures that
the same versions of all dependencies are installed during successive
builds. The environment variables with package information are removed
since we no longer combine them and to pass Hadolint linting (this is
manual and not included in our automatic linting).