Use the ansible:3.17 Docker image for the compile-stage
Switch to using the ansible:3.17 Docker image for the compile-stage
since the python3-dev package will install the python3 package and
this results in two different Python installs in the compile-stage. We
instead keep the python Docker image used in the build-stage on the
same version of Python as is available in Alpine Linux 3.17 to ensure
compatibility with the created venv.