Selectively use git CLI mode with cargo
Only create the cargo configuration file with the setting to use the
git CLI if the target architecture is not amd64 or arm64. This is
especially relevant Since these two platforms are natively supported by
the cryptography package so they should not need to build wheels in the
first place.