Use a static Cargo configuration file
Instead of creating a Cargo configuration file in the Dockerfile we
instead store an appropriate Cargo configuration file and COPY it into
the image. This allows the file to be checked by pre-commit and
appropriately version controlled. It also makes it easier to make
additional configuration changes from default if needed.