Update pipenv package installation
Add `pipenv check` to ensure that any PEP 508 requirements are met and
to check for any security vulnerabilities in the lock file
configuration. Replace `pipenv sync` with `pipenv install --deploy`.
They will both install all of the packages specified in the
Pipfile/Pipfile.lock files but the new method will ensure that the
Pipfile and Pipfile.lock files are in sync.