Use the `--break-system-packages` when installing `pipenv`
Alpine Linux 3.20 configures Python 3 as externally managed, so we need
to pass this flag to `pip` to install a Python package directly. Since we
are using this to build a Python virtual environment in the
`compile-stage` that is moved to the `build-stage` for use in the final
image this should pose no issues.