Use the system package version of the `cryptography` Python package
I am having timeout issues when building a wheel for the `cryptography`
package on platforms that do not have a wheel available from PyPI.
Switching to a system package version is the sanest (to me) workaround.
The system packages that are installed to build the package on these
platforms are removed as part of this change.