Bump Alpine Linux from 3.20 to 3.21
The version of the `py3-cryptography` package available in Alpine
Linux 3.20 is vulnerable to PVE-2024-73711 and CVE-2024-4603. As a
result `pipenv` is unable to install the package dependencies for the
project. Moving to Alpine Linux 3.21 allows us to use a newer version
of the `py3-cryptography` package that is not vulnerable.